lakefs-test/image-segmentation-local/requirements.txt
# Basic ML/DL libraries
torch>=1.9.0
torchvision>=0.10.0

# Image processing
Pillow>=8.0.0
numpy>=1.20.0

# MLflow
mlflow>=2.8.0

# lakeFS
lakefs-client>=1.0.0 

# Other utilities
python-dotenv>=0.19.0
GitPython>=3.1.43